refactor(test): 패키지 분리 해제
패키지를 원래대로 되돌렸습니다.
