+++ src/test/java/lotto/MyLottoTest.java
@@ -1,5 +1,6 @@
-package model;
+package lotto;
 
+import model.MyLotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
